Update sonicfight.html
Implemented CSP to prevent vulnerabilities with the iframe on the page. For the next iframe (that doesn't need JavaScript), I'll sandbox it and implement CSP to boot :)

People need security, and it is here. IFrames may be convenient, but only use them when you want to create a gameplay page, like I did, or if you need to with no other alternative :)